requirements.txt
streamlit
pandas
plotly-express
xlsxwriter